+++ src/main/java/codesquad/fineants/spring/api/member/aop/HasNotificationAuthorizationAspect.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.auth;
+package codesquad.fineants.spring.api.member.aop;
 
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
