+++ src/test/java/codesquad/fineants/spring/api/portfolio/PortFolioRestControllerTest.java
@@ -39,6 +39,7 @@ import codesquad.fineants.domain.oauth.support.AuthPrincipalArgumentResolver;
 import codesquad.fineants.domain.portfolio.Portfolio;
 import codesquad.fineants.domain.portfolio.PortfolioRepository;
 import codesquad.fineants.spring.api.common.errors.handler.GlobalExceptionHandler;
+import codesquad.fineants.spring.api.portfolio.aop.HasPortfolioAuthorizationAspect;
 import codesquad.fineants.spring.api.portfolio.controller.PortFolioRestController;
 import codesquad.fineants.spring.api.portfolio.request.PortfolioCreateRequest;
 import codesquad.fineants.spring.api.portfolio.request.PortfolioModifyRequest;
@@ -47,7 +48,6 @@ import codesquad.fineants.spring.api.portfolio.response.PortFolioItem;
 import codesquad.fineants.spring.api.portfolio.response.PortfolioModifyResponse;
 import codesquad.fineants.spring.api.portfolio.response.PortfoliosResponse;
 import codesquad.fineants.spring.api.portfolio.service.PortFolioService;
-import codesquad.fineants.spring.auth.HasPortfolioAuthorizationAspect;
 import codesquad.fineants.spring.config.JpaAuditingConfiguration;
 import codesquad.fineants.spring.config.SpringConfig;
 import codesquad.fineants.spring.util.ObjectMapperUtil;
